Moved from symbolic-pcode repository (#1)
Code moved from symbolic-pcode repository. Since it was previously in a
workspace the `Cargo.toml` file needed to be updated to reflect those
changes, but otherwise no changes have been made to source code.

The original integration tests have been merged into the unit tests in
this repository. They were originally factored out due to reliance on
building and running a local version of the SLEIGH compiler. Now that
the `sleigh-config` crate is a dependency this separation is no longer
necessary.