Remove XML SLA decoding from GhidraSleighBuilder (#6)
The XML data passed down into libsla-sys is a wrapper that directs the
underlying libsla library to decode a `.sla` file. In addition to being
an unnecessary trampoline with no value, it limits the interface by
requiring that the sla data reside in a file on disk. Maintaining this
behavior provides no value and as such is being removed.