Improve disassembly ergonomics (#7)
Internally the native and pcode disassembly went through the same code
path. This resulted in a native disassembly which supposedly might
disassemble multiple instructions when in fact it would only ever
disassemble one. The code has been refactored so that the native
disassembly type reflects this reality, no longer requiring callers to
pull off the first (and only) instruction.