Use AddressSpace name instead of id to map to sys::AddrSpace (#12)
The AddressSpace id changes between Sleigh invocations. This id was used
for mapping address spaces since pcode instructions use this id for
identifying address spaces. However there's no requirement for Sleigh to
align with pcode here, and the address space name is a better suited
identifier due to it being a fixed value.

Resolves #11